Add initial files
Added some starter files. Mining.html is the runner file. Main.js contains the part of the game that runs it all. Enums and Config are files that store data about the blocks and the game configuration. Block is a constructor for the blocks you mine. Player is the player object. Control handles the player's controls.